refactor: split Settings.tsx component into Logout.tsx, Themes.tsx and UpdateProfile.tsx components
